Split configuration structures into a separate file
